fix bad code
- Fix improperly calling a server script func without being in the server frame (from tracker main tasks thread)

- Remove terrible implementation of removing a ref, since it doesn't even check for garbage colleciton and Release already performs this

- Release ref in server frame

This entire process should just send up a table from the stack, and be using a more centralized json format with potentially using web socket protocol or if still using curl for api responses, change the format of return, build and push a table onto the stack, change the return type to accept table instead of string -- repeat the process with PlayerData functions.

- Todo: simplify tracker to fetch squirrel objects that hold stats instead of using legacy log entry formatted string copies, or just replace with liveapi for outbound data, build custom plugin for inbound data? Lots of work. It would be more practical to rebuild the script version of tracker stats to only hold a unified data structure for stats.  When this tracker project started, fight history and overall stat recording was broken and convoluted in scripts, but now may be feasible to hook into the games systems accurately, especially with the extensive work done in the sdk and squirrel api.  This would require backend refactors at processing classes to accept new json format over custom format and a massive refactor.

For now this system is functional and allows data to be processed until  enough allocated time is available to create a uniform and proper piece of production software.